requirements.txt
supervision
ultralytics
openvino
opencv-python
flask
gunicorn
flask-mqtt
orjson
numpy
pillow